+++ R/crew_controller_sge.R
@@ -1,7 +1,7 @@
 #' @title `r lifecycle::badge("maturing")` Create a controller with a
 #'   Sun Grid Engine (SGE) launcher.
 #' @export
-#' @family plugin_sge
+#' @family sge
 #' @description Create an `R6` object to submit tasks and
 #'   launch workers on Sun Grid Engine (SGE) workers.
 #' @inheritSection crew.cluster-package Attribution