+++ R/crew_launcher_slurm.R
@@ -1,7 +1,7 @@
 #' @title `r lifecycle::badge('experimental')` Create a launcher with
 #'   SLURM workers.
 #' @export
-#' @family plugin_slurm
+#' @family slurm
 #' @description Create an `R6` object to launch and maintain
 #'   workers as SLURM jobs.
 #' @details WARNING: the `crew.cluster` SLURM plugin is experimental
@@ -125,7 +125,7 @@ crew_launcher_slurm <- function(
 
 #' @title `r lifecycle::badge("experimental")` SLURM launcher class
 #' @export
-#' @family plugin_slurm
+#' @family slurm
 #' @description `R6` class to launch and manage SLURM workers.
 #' @details See [crew_launcher_slurm()].
 #' @inheritSection crew.cluster-package Attribution