+++ R/crew_monitor_sge.R
@@ -1,10 +1,8 @@
 #' @title `r lifecycle::badge("experimental")` Create a SGE monitor object.
 #' @export
-#' @family monitor
+#' @family sge
 #' @description Create an `R6` object to monitor SGE cluster jobs.
-#' @inheritParams crew_launcher_cluster
-#' @param command_list Character of length 1,
-#'   file path to the executable to list jobs.
+#' @inheritParams crew_monitor_cluster
 crew_monitor_sge <- function(
   verbose = TRUE,
   command_list = as.character(Sys.which("qstat")),
@@ -21,65 +19,14 @@ crew_monitor_sge <- function(
 
 #' @title `r lifecycle::badge("experimental")` SGE monitor class
 #' @export
-#' @family monitor
+#' @family sge
 #' @description SGE monitor `R6` class
 #' @details See [crew_monitor_sge()].
 crew_class_monitor_sge <- R6::R6Class(
   classname = "crew_class_monitor_sge",
+  inherit = crew_class_monitor_cluster,
   cloneable = FALSE,
-  private = list(
-    .verbose = NULL,
-    .command_list = NULL,
-    .command_terminate = NULL
-  ),
-  active = list(
-    #' @field verbose See [crew_monitor_sge()].
-    verbose = function() {
-      .subset2(private, ".verbose")
-    },
-    #' @field command_list See [crew_monitor_sge()].
-    command_list = function() {
-      .subset2(private, ".command_list")
-    },
-    #' @field command_terminate See [crew_monitor_sge()].
-    command_terminate = function() {
-      .subset2(private, ".command_terminate")
-    }
-  ),
   public = list(
-    #' @description SGE monitor constructor.
-    #' @return an SGE monitor object.
-    #' @param verbose See [crew_monitor_sge()].
-    #' @param command_list See [crew_monitor_sge()].
-    #' @param command_terminate See [crew_monitor_sge()].
-    initialize = function(
-      verbose = NULL,
-      command_list = NULL,
-      command_terminate = NULL
-    ) {
-      private$.verbose <- verbose
-      private$.command_list <- command_list
-      private$.command_terminate <- command_terminate
-    },
-    #' @description Validate the monitor.
-    #' @return `NULL` (invisibly).
-    validate = function() {
-      crew::crew_assert(
-        private$.verbose,
-        isTRUE(.) || isFALSE(.),
-        message = "'verbose' must be TRUE or FALSE."
-      )
-      for (field in c(".command_terminate", ".command_list")) {
-        crew::crew_assert(
-          private[[field]],
-          is.character(.),
-          length(.) == 1L,
-          !anyNA(.),
-          message = paste(field, "must be a valid length-1 character string.")
-        )
-      }
-      invisible()
-    },
     #' @description List SGE jobs.
     #' @return A `tibble` with one row per SGE job and columns with
     #'   specific details.