+++ NEWS.md
@@ -1,6 +1,6 @@
-# crew.cluster 0.2.0.9000 (development)
+# crew.cluster 0.2.0.9001 (development)
 
-
+* Create an abstract monitor class for cluster-specific monitor classes to inherit from (#32).
 
 # crew.cluster 0.2.0